!!! COMPILE PHASE_3 !!! Reward Panel changes: Added reward panel to the end of cog office battles. Added reward panel to the end of the tutorial. Hid +0 labels next to tracks that aren't unlocked. Solidified reward panel system code from my last commit.
Backpack changes:
Changed default backpack to only the cupcake and the squirting flower.
Temporarily "buffed" the cupcake to do 50 damage and have a max carry of 30.
Removed "auto-update" code that added missing gags to backpacks in #announceGenerate() of DToonAI.

Misc changes:
Added in two test quests.
The Gag Shop now has supports for showing <4 gags in a frame.
More progress on the inspect quest objective.


Former-commit-id: b58c76073cae5c93e8533bc9d76a227133ca1868