Add admin buttons for unlocking all gags and unlocking all gag slots. Add stun effect over Cog head when killed. Add an attack cooldown based on the cog's level (so they don't spam attack as much). Add a bit of lock-on for GlowerPower.
Reduce speed of ThrowAttacks (also, a better system for when to release the projectile).
Make cogs have a 20% chance of taunting when using an attack (it got annoying).

Fix incorrect camera angles for the victory sequence.
Fix too much lock-on for FountainPen.
Fix district reset when killing a take-over suit.

Some other tweaks.


Former-commit-id: 7fbcd9d8e8203734a4fd755d1f50cba1290b4573